backend/requirements.txt
flask
flask-cors
bcrypt
transformers 
accelerate
torch
llama-cpp-python
